HCI and Psychology
HCI and Psychology – an introduction

Interaction with a computer is the same as interaction with information
which involves processing of information in the brain which in turn
involves cognitive processes. So, cognitive psychology and system design
are interlinked to ensure that this information processing activity is
within users' capabilities.
Cognitive psychology can be applied to improve system design:-
to provide knowledge of user's capabilities
to identify and explain user problems
to supply modelling tools

You Need a Psychologist to Teach HCI Correctly to a Computer Scientist
For the first time since its inauguration, IFIP's conference on human
computer interaction, INTERACT, made its way to the southern hemisphere.
INTERACT'97 was held in Sydney, Australia, in July, and incorporated two
regional conferences, OZCHI97 and APCHI97. The theme of the conference
was "Discovering New Worlds of HCI" which drew participants from all
over the globe and from a wide range of disciplines.

It was this range of disciplines which set the scene for a lively
discussion in the opening panel session where the proposition was put
forward that a psychologist is needed to teach HCI correctly to computer
scientists.

If HCI is to be successfully adopted and practised in industry, software
developers will require some form of education or training in this area.
The question remains as to who should be providing this teaching -- a
psychologist, or a computer scientist? The panel and audience grappled
with a number of questions during the course of the session. As well as
the question of exactly who should be responsible for the teaching of
material, other related issues were discussed including who should be
responsible for the content, what material should be presented, and how
it should be taught.

In attempting to address the main question, it was not surprising that
the related question of what needs to be taught to these individuals was
raised a number of times. While it is not possible to teach computer
science students everything they will need to know about HCI, it was
recognised that a key objective is to increase the awareness of students
of their own limitations and the boundaries of their discipline so that
they learn to ask important questions and raise relevant issues during
the course of their practice. Ideally, the student should be taught to
draw from the knowledge provided by external disciplines and be able to
apply it to the problem at hand. The material taught needs to be
well-focussed and provide students with a strong sense of what HCI is
about and what it tries to achieve. It needs to be directly relevant to
the capabilities and needs of the students to whom it is presented, and
able to be used in industry both currently and in the long term.

Computer science students by their nature were recognised as having a
rigid approach in their thinking and reasoning. This can give rise to
incompatibilities between the more qualitative nature of psychological
knowledge and the more quantitative heuristics needed by software
developers. While computer scientists may be able to communicate to
their students at this level, there is a danger that they may misapply
psychological knowledge or oversimplify human behaviour. An example was
given where a computer scientist had taught a class that graphical user
interfaces were bad when they had more than seven items which was
thought to be in conflict with Miller's recommendation for short-term
memory.

It was pointed out by both sides that no single person is likely to be
able to give a comprehensive picture of HCI as a whole. Computer
scientists are likely to have a poor understanding of areas in the
social sciences. Psychologists are likely to have a poor understanding
of the software engineering context required to build systems. Even
multi-disciplinary academics and practitioners can have their own
specific prejudices due to their limited perspectives on HCI. So where
does the solution lie? Two general approaches were offered. The first
makes use of a team of individual specialists to provide a broad scope.
The second recognises HCI as a discipline in its own right which needs
to be taught by those who understand HCI as a whole, rather than just
specific components. However, neither of these approaches were seen as
being particularly easy to implement.

Finally, the question of how the material should be taught was also
posed. The notion of trying to integrate the teaching and assessment of
HCI components throughout the entire curriculum rather than just in
isolated units was supported. For example, a number of marks could be
specifically allocated towards usability in programming assignments. In
addition, there was a recognised need to place students in large and
realistic projects, possibly as a result of partnerships between
industry and universities.

There appears to be significant parallels between HCI education and the
HCI discipline as a whole. Firstly, as HCI struggles to define itself,
HCI education needs to focus its scope and objectives. Secondly, as
industry and academia look to match results and needs, students need to
distinguish between what HCI can do, and what it does in practice.
Thirdly, as each of the HCI component disciplines search for common
ground, students need to be taught to recognise the boundaries and be
able to apply the relevant information to the problem at hand. With such
high goals, perhaps a more pertinent question is, "Who is the most
appropriate person to teach HCI to the teachers?"

Panel Organiser: Marilyn Mantei, Canada
Chair: Penny Collings, Australia
Panelists: Tom Hewett, USA; Andrew Monk, UK; Lorraine Johnston,
Australia; Mark Apperley, New Zealand

Paul Chesson is completing a Ph.D. at the University of Melbourne,
Australia, in the area of formal methods in HCI. His research involves
the development of a dialogue specification language for interactive
systems, and a supporting method and tools for its use. He holds a B.E.
from Melbourne in software engineering. Email: chesson@cs.mu.oz.au